requirements.txt
aio-pika==9.5.5
aiormq==6.8.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
clerk-backend-api==1.8.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
dnspython==2.7.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.11
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
kombu==5.5.0
motor==3.7.0
multidict==6.2.0
mypy-extensions==1.0.0
neo4j==5.28.1
pamqp==3.3.0
prompt_toolkit==3.0.50
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
